Purpose: Add feature to include a player's name in the console output of ServerLink
Overview of changes needed:

Prior to these changes, when a Player joins a UT2004 server, the 3SPN
calls in to ServerLink with "GET_STATS" which calls the statGetPlayerStat
function.  At this point ServerLink statGetPlayerStat function does
NOT know the player name.  It is only aware of player hash from 3SPN.
ServerLink does however cash the DB to memory in PlayerAggregate on
startup, but (prior to this change) it does NOT have a player name
value anywhere in its cache.

Side Effects: None at this time.

Detailed Changes on per file basis:

ServerLinkHandler.java:
-  Comment out and add code so that internally we can print all
inbound 3SPN requests for dev purposes.

-  Add "Player name" and mLoggerPlayerName value from cached
internal data store to ServerLink's std output as a result
of a call to new function PlayerAggregate.getLoggerPlayerName.

PlayerAggregate.java:
- Add new member var mLoggerPlayerName.  This never gets written
to the database.  It is only for purposes of logging the username
to the console.
- Add getter and setter function's getLoggerPlayerName and
setLoggerPlayerName.

SQLiteDatabaseRepositoryImpl.java:
- Comment out and replace db read query statement to include
MAX(timestamp) - so we get the player's most recent in game
player name; without this we would only acquire the first
player name they used from the DB.  Add playerName to the
new query statement too.
- Make call to setLoggerPlayerName in the open method so as to
obtain and store the player's name to our internal cached data
store on startup db open.
- Make call to setLoggerPlayerName in the putPlayerStat method
so as to update our internal cache of the player name's in
case anyone changed their name(s) during the match.  Results
will be seen on the next match when 3SPN calls in with "GET_STATS".